requirements.txt
# SCROLLS RAG Baseline Benchmark — Python Dependencies
# Install: pip install -r requirements.txt

# Core ML
torch>=2.1.0
transformers>=4.40.0
sentence-transformers>=2.7.0
datasets>=2.19.0

# Vector search
faiss-cpu>=1.7.4
# If GPUs available and you want GPU-accelerated FAISS, swap to:
# faiss-gpu>=1.7.4

# LLM inference (local vLLM)
vllm>=0.4.0

# Evaluation
rouge-score>=0.1.2
nltk>=3.8.0

# Utilities
numpy>=1.24.0
tqdm>=4.65.0
